Initial UI for Rock Paper Scissor Game.
Create three buttons for user input
Create Scorecard
Create initial game logic on Rock Button. An event listener was added
upon clicking the Rock Button which triggers the game updating the Scorecard
variable where the results are stored

On the the next commit the logic will be refactored so it also takes
into account the other buttons: paper and scissor.